Remove simple-bloom gossip type
