+++ crates/kitsune_p2p/kitsune_p2p/src/gossip.rs
@@ -1,4 +1,6 @@
 //! Various gossip strategies for kitsune
 
 pub mod sharded_gossip;
-pub mod simple_bloom;
+
+mod common;
+pub use common::*;
