+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/space.rs
@@ -1127,10 +1127,6 @@ impl Space {
             .gossip_strategy
             .split(',')
             .flat_map(|module| match module {
-                "simple-bloom" => vec![(
-                    GossipModuleType::Simple,
-                    crate::gossip::simple_bloom::factory(),
-                )],
                 "sharded-gossip" => vec![
                     (
                         GossipModuleType::ShardedRecent,
